re-arranged symbolic directory to make things clearer
