+++ sysopt/symbolic/__init__.py
@@ -5,3 +5,5 @@
 
 from sysopt.symbolic.symbols import *
 from sysopt.symbolic.decision_variables import *
+from sysopt.symbolic.numpy_api import *
+from sysopt.symbolic.graph_evaluators import *
